fix in train and test

+++ src/test.py
@@ -14,6 +14,6 @@ def test(model, test_loader, k=1, device=torch.device("cpu"),):
             update = prediction.eq(labels.view(-1, 1).expand_as(prediction))
             correct += update.sum().item()
             total += images.shape[0]
-    accuracy = correct.float() / total
+    accuracy = correct / total
     return accuracy
